ls for mac

+++ setup.sh
@@ -1,15 +1,21 @@
 #! /usr/bin/env bash
 
-CONFIG_DIR=$HOME/.linux-terminal/
+CONFIG_DIR=$HOME/.linux-terminal
 VIM_DIR=$HOME/.vim/
 VIMRC=$HOME/.vimrc
-ROOT=/root
+
+
+if [ $TERM_PROGRAM == "Apple_Terminal" ]; then
+    is_mac=1 
+else 
+    is_mac=0 
+fi
 
 
 progress=1
 
 function echo_progress {
-    printf "$progress. \e[1;33m%.0s---> \e$1\n"
+    printf "\e[1;36m$progress. %.0s---> \e[1;32m$1\e[0m\n"
     (( progress = $progress + 1 ))
 }
 
@@ -23,7 +29,7 @@ elif [ -e /usr/bin/yum ]; then
     sudo yum install http://pkgs.repoforge.org/htop/htop-1.0.1-2.el6.rf.x86_64.rpm
     sudo easy_install pip
     sudo pip install grc
-elif [ $TERM_PROGRAM == "Apple_Terminal" ]; then 
+elif [[ $is_mac ]]; then 
     ruby -e "$(curl -fsSkL raw.github.com/mxcl/homebrew/go)"
     brew install git bash-completion vim mc htop
     brew install python --universal
@@ -44,7 +50,7 @@ echo_progress "Configuring vim ..."
 sudo chown -R $USER $HOME/.vim/
 rm -rf $VIM_DIR
 git clone git://github.com/pavelulyashev/vim.git $VIM_DIR
-echo_progress "source $HOME/.vim/rc.vim" > $VIMRC
+echo "source $HOME/.vim/rc.vim" > $VIMRC
 sudo chown -R $USER $HOME/.vim/
 cd $VIM_DIR
 git submodule init && git submodule update
@@ -56,29 +62,37 @@ if [ -d $HOME/.mc ]; then
 elif [ -d $HOME/.config/mc ]; then
     cp $CONFIG_DIR/mc_ini $HOME/.config/mc/ini
 fi
-sudo mkdir $ROOT/.mc
-sudo ln -sf $HOME/.mc/ini $ROOT/.mc/ini
 
 # moc
-echo_progress "Configuring moc ..."
-mkdir $HOME/.moc
-cp $CONFIG_DIR/moc_config $HOME/.moc/config
-chmod 644 $HOME/.moc/config
+if [ -f /usr/bin/mocp ]; then
+    echo_progress "Configuring moc ..."
+    [ ! -d $HOME/.moc ] && mkdir $HOME/.moc
+    cp $CONFIG_DIR/moc_config $HOME/.moc/config
+    chmod 644 $HOME/.moc/config
+fi
 
 # screen
 echo_progress "Configuring screen ..."
 cp $CONFIG_DIR/.screenrc $HOME
-sudo ln -sf $HOME/.screenrc $ROOT/.screenrc
 
 # git
 echo_progress "Configuring git ..."
 cp $CONFIG_DIR/.gitconfig $HOME
-sudo ln -sf $HOME/.gitconfig $ROOT/.gitconfig
 
 # aliases
 echo_progress "Configuring root bash ..."
-cp $CONFIG_DIR/bash/.bash* $HOME
-for file in $(ls -a $CONFIG_DIR/bash | grep .bash); do sudo ln -sf $HOME/$file $ROOT/$file; done
+if [[ ! $is_mac ]]; then
+    ROOT=/root
+
+    sudo ln -sf $HOME/.gitconfig $ROOT/.gitconfig
+    sudo ln -sf $HOME/.screenrc $ROOT/.screenrc
+
+    cp $CONFIG_DIR/bash/.bash* $HOME
+    for file in $(ls -a $CONFIG_DIR/bash | grep .bash); do sudo ln -sf $HOME/$file $ROOT/$file; done
+
+    sudo mkdir $ROOT/.mc
+    sudo ln -sf $HOME/.mc/ini $ROOT/.mc/ini
+fi
 
 # install mysql + copy grc config
